add addLLMPanelListeners
