+++ src/main/kotlin/org/jetbrains/research/testspark/actions/llm/LLMParametersUtils.kt
@@ -4,10 +4,12 @@ import com.google.gson.JsonParser
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.ui.ComboBox
 import com.intellij.util.io.HttpRequests
-import org.jetbrains.research.testspark.settings.SettingsApplicationState
+import org.jetbrains.research.testspark.services.SettingsApplicationService
 import java.net.HttpURLConnection
 import javax.swing.DefaultComboBoxModel
 import javax.swing.JTextField
+import javax.swing.event.DocumentEvent
+import javax.swing.event.DocumentListener
 
 /**
  * Checks if the Grazie class is loaded.
@@ -24,6 +26,65 @@ fun isGrazieClassLoaded(): Boolean {
 }
 
 /**
+ * Adds listeners to the given components to handle events and perform necessary actions.
+ *
+ * @param platformSelector The combo box used for selecting platforms.
+ * @param modelSelector The combo box used for selecting models.
+ * @param llmUserTokenField The text field used for entering the user token.
+ * @param defaultModulesArray An array of default module names.
+ * @param lastChosenModule The name of the last chosen module.
+ */
+fun addLLMPanelListeners(
+    platformSelector: ComboBox<String>,
+    modelSelector: ComboBox<String>,
+    llmUserTokenField: JTextField,
+    defaultModulesArray: Array<String>,
+    lastChosenModule: String,
+) {
+    llmUserTokenField.document.addDocumentListener(object : DocumentListener {
+        override fun insertUpdate(e: DocumentEvent?) {
+            updateModelSelector(
+                platformSelector,
+                modelSelector,
+                llmUserTokenField,
+                defaultModulesArray,
+                lastChosenModule,
+            )
+        }
+
+        override fun removeUpdate(e: DocumentEvent?) {
+            updateModelSelector(
+                platformSelector,
+                modelSelector,
+                llmUserTokenField,
+                defaultModulesArray,
+                lastChosenModule,
+            )
+        }
+
+        override fun changedUpdate(e: DocumentEvent?) {
+            updateModelSelector(
+                platformSelector,
+                modelSelector,
+                llmUserTokenField,
+                defaultModulesArray,
+                lastChosenModule,
+            )
+        }
+    })
+
+    platformSelector.addItemListener {
+        updateModelSelector(
+            platformSelector,
+            modelSelector,
+            llmUserTokenField,
+            defaultModulesArray,
+            lastChosenModule,
+        )
+    }
+}
+
+/**
  * Updates the model selector based on the selected platform in the platform selector.
  * If the selected platform is "Grazie", the model selector is disabled and set to display only "GPT-4".
  * If the selected platform is not "Grazie", the model selector is updated with the available modules fetched asynchronously using llmUserTokenField and enables the okLlmButton.
@@ -35,10 +96,11 @@ fun updateModelSelector(
     platformSelector: ComboBox<String>,
     modelSelector: ComboBox<String>,
     llmUserTokenField: JTextField,
-    settingsState: SettingsApplicationState,
     defaultModulesArray: Array<String>,
     lastChosenModule: String,
 ) {
+    val settingsState = SettingsApplicationService.getInstance().state!!
+
     if (platformSelector.selectedItem!!.toString() == "Grazie") {
         modelSelector.model = DefaultComboBoxModel(arrayOf("GPT-4"))
         modelSelector.isEnabled = false
