+++ src/main/kotlin/org/jetbrains/research/testspark/settings/evosuite/SettingsEvoSuiteConfigurable.kt
@@ -1,9 +1,10 @@
-package org.jetbrains.research.testspark.settings
+package org.jetbrains.research.testspark.settings.evosuite
 
 import com.intellij.openapi.options.Configurable
 import com.intellij.openapi.ui.Messages
 import org.jetbrains.research.testspark.services.SettingsApplicationService
 import javax.swing.JComponent
+import org.jetbrains.research.testspark.settings.SettingsApplicationState
 
 /**
  * This class allows to configure some EvoSuite settings via the EvoSuite page in the Settings dialog,
